Move most tests to source files using InlineTest.jl
