+++ src/Vision/recipes.jl
@@ -142,3 +142,63 @@ function _registerrecipes()
         Datasets.registerrecipe!(Datasets.FASTAI_DATA_REGISTRY, name, recipe)
     end
 end
+
+
+# ## Tests
+
+
+@testset "ImageFolders [recipe]" begin
+    path = joinpath(datasetpath("mnist_var_size_tiny"), "train")
+
+    @testset "Basic configuration" begin
+        recipe = Vision.ImageFolders()
+        data, blocks = loadrecipe(recipe, path)
+        Datasets.testrecipe(recipe, data, blocks)
+        @test blocks[1] isa Image
+        @test blocks[2].classes == ["3", "7"]
+    end
+
+    @testset "Split configuration" begin
+        recipe = Vision.ImageFolders(split=true)
+        data, blocks = loadrecipe(recipe, path)
+        Datasets.testrecipe(recipe, data["train"], blocks)
+    end
+
+    @testset "Error cases" begin
+        @testset "Empty directory" begin
+            recipe = Vision.ImageFolders(split=true)
+            @test_throws ErrorException loadrecipe(recipe, mktempdir())
+        end
+
+        @testset "Only one label" begin
+            recipe = Vision.ImageFolders(labelfn=x -> "1")
+            @test_throws ErrorException loadrecipe(recipe, path)
+        end
+    end
+
+end
+
+
+@testset "ImageSegmentationFolders [recipe]" begin
+    path = datasetpath("camvid_tiny")
+
+    @testset "Basic configuration" begin
+        recipe = ImageSegmentationFolders()
+        data, blocks = loadrecipe(recipe, path)
+        Datasets.testrecipe(recipe, data, blocks)
+        @test blocks[1] isa Image
+        @test blocks[2] isa Mask
+    end
+
+    @testset "Error cases" begin
+        @testset "Empty directory" begin
+            recipe = ImageSegmentationFolders()
+            @test_throws ErrorException loadrecipe(recipe, mktempdir())
+        end
+
+        @testset "Only one label" begin
+            recipe = ImageSegmentationFolders(labelfile="idontexist")
+            @test_throws ErrorException loadrecipe(recipe, path)
+        end
+    end
+end
