+++ src/datasets/registry.jl
@@ -133,3 +133,43 @@ function blocksmatch(recipe::DatasetRecipe, TBlocks)
     @show recipeblocks(recipe), TBlocks
 	return recipeblocks(recipe) <: TBlocks
 end
+
+
+# ## Tests
+
+
+
+@testset "DatasetRegistry" begin
+    reg = Datasets.DatasetRegistry()
+    @testset "registerdataset!" begin
+        @test_nowarn Datasets.registerdataset!(
+            reg, "mnist_var_size_tiny", () -> datasetpath("mnist_var_size_tiny"))
+
+
+        # Reregistering should error
+        @test_throws ErrorException Datasets.registerdataset!(
+            reg, "mnist_var_size_tiny", () -> datasetpath("mnist_var_size_tiny"))
+    end
+
+    @testset "listdatasources" begin
+        @test length(Datasets.listdatasources(reg)) == 1
+    end
+
+    @testset "datasetpath" begin
+        @test_nowarn datasetpath(reg, "mnist_var_size_tiny")
+    end
+
+    @testset "registerrecipe!" begin
+        @test_nowarn Datasets.registerrecipe!(
+            reg, "mnist_var_size_tiny", Vision.ImageFolders())
+    end
+
+    @testset "finddatasets" begin
+        @test finddatasets(reg) |> length == 1
+        @test finddatasets(reg, name="mnist_var_size_tiny") |> length == 1
+        @test finddatasets(reg, blocks=Tuple{Image, Label}) |> length == 1
+        @test finddatasets(reg, blocks=Tuple{Image, LabelMulti}) |> length == 0
+        @test finddatasets(reg, name="mnist_var_size_tiny", blocks=Tuple{Image, Label}) |> length == 1
+        @test finddatasets(reg, name="mnist", blocks=Tuple{Image, Label}) |> length == 0
+    end
+end
