--- test/encodings/projective.jl
@@ -1,1 +1,0 @@
-include("../imports.jl")
