+++ test/fasterai.jl
@@ -2,7 +2,7 @@ include("imports.jl")
 
 
 
-@testset ExtendedTestSet "FasterAI" begin
+@testset "FasterAI" begin
     @test length(listdatasources()) > 10
 
     @test !isempty(finddatasets(blocks=(Image, Label)))
