+++ test/training/finetune.jl
@@ -1,4 +1,3 @@
-include("../imports.jl")
 
 
 struct ConstGrouper <: FastAI.ParamGrouper
@@ -6,7 +5,7 @@ struct ConstGrouper <: FastAI.ParamGrouper
 end
 FastAI.group(cg::ConstGrouper, m) = Dict(cg.g => m)
 
-@testset ExtendedTestSet "`finetune!`" begin
+@testset "finetune!" begin
     learner = testlearner(Recorder())
     @test_nowarn finetune!(learner, 1; grouper = ConstGrouper(2))
