+++ test/training/lrfind.jl
@@ -1,7 +1,4 @@
-include("../imports.jl")
-
-@testset ExtendedTestSet "lrfind" begin
+@testset "lrfind" begin
     learner = testlearner()
     @test_nowarn result = lrfind(learner)
-
 end
