+++ test/training/paramgroups.jl
@@ -1,10 +1,9 @@
-include("../imports.jl")
 
-@testset ExtendedTestSet "`ParamGroups`" begin
+@testset "ParamGroups" begin
     model = Chain(Dense(3, 5), Dense(5, 3))
-    paramgroups = ParamGroups(IndexGrouper([1, 2]), model)
+    paramgroups = FastAI.ParamGroups(FastAI.IndexGrouper([1, 2]), model)
 
-    @test getgroup(paramgroups, model[1].weight) == 1
-    @test getgroup(paramgroups, model[2].weight) == 2
-    @test getgroup(paramgroups, rand(10)) === nothing
+    @test FastAI.getgroup(paramgroups, model[1].weight) == 1
+    @test FastAI.getgroup(paramgroups, model[2].weight) == 2
+    @test FastAI.getgroup(paramgroups, rand(10)) === nothing
 end
